Basics of GH and encryption integration.
V2 app has the skeleton of the app, as well as a rather large block of RX
logic for establishing the key file in the repo.

Sets up the file and does a challenge if the file already exists.

Challenge is in the form of the salt encrypted using the public key, so it
needs comparing for equality to the salt pulled from the key file after
decrypting the private key using the password entered by the user.

This lays down the basic support of having a cryptographic identity tied
to users of the application.

Might be overkill, but RSA at 3072 and Aes at 256 respectively for key
size, doesn't seem to have a noticeably poor effect on performance, and it
only uses about 3.5kb in GH. So might be worth the extra protection. Of
course, it'll translate to inflating the image somewhat by adding in
signatures, but that's fine.

Next up is the proof-of-concept of packing the image and putting the ACL
file into the pack-hub repo for referencing.

An end-to-end secure transfer proof is needed for confidence.